Port Mossy Mayhem's wall lattice to C#, and widen IScenery to long
MossWalls: the unbounded 64-unit cell lattice, generated on demand from
(seed, block coordinate) and memoized with FIFO eviction. Block dealing
(the xorshift32 BlockRng, the four shape kinds — line/L/T/room — with
their inset margins), collision (overlap, destructible overlap, the
multi-pass corner-aware push-out with its buried-body fallback), DDA
raycasting, and the tree-stem damage model are all here.

Driven entirely through the public API, because blockKey/BLOCK_CELLS/
generateBlock aren't exported from the TypeScript module and there's no
way to pose a block directly without editing wallsMossy.ts itself. A
dense WallKindAt sweep across five seeds stands in for a raw block dump —
every cell it names has already been through the hash, the block RNG and
the shape dealer exactly as a real query would, so it's the generation-
determinism check as much as a query check. The push-out property sweep
mirrors tests/wallsMossy.test.ts's own 100,000-probe invariant check
exactly, so no oracle is needed there — both languages just have to
satisfy the same "never leaves a body inside a wall" property.

Found a real, first-draft bug: PackWallCell's (cx+2^20)*2^21+(cy+2^20)
packing overflows int32 for EVERY real cell, not just extreme ones — the
2^20*2^21 term alone is ~2.2e12, about a thousand times past int32's
~2.1e9 range. JavaScript never notices (a plain number is exact to 2^53);
C#'s `int` silently wrapped. Caught by the fixture on the very first C#
run, before any fault was deliberately injected — every overlap and ray
probe failed with values like -112197691 instead of 2198911057861.

The fix reaches further than this file: IScenery's index-typed members
(Overlap, DestructibleOverlap, Damage, Destroy, IsDestructible, PieceX,
PieceY, PieceRadius) are now `long`, not `int`, on the shared interface —
a wall lattice's packed cell coordinate needs the range, even though
ScrapPiles' own dense-array indices stay small and are narrowed back
internally the moment they touch its arrays. EnemyAI.cs's two scenery-
index locals (`ahead`, `beside`) widen to match. Nothing that only ever
compared an index against zero needed to change, and all 105 pre-existing
C# tests still pass unmodified.

One genuinely unreachable branch, found and recorded rather than forced:
push-out's "buried with literally no open cardinal face" fallback cannot
be reached by anything this generator draws — every shape is one cell
thick, so every occupied cell has at least two empty cardinal neighbours
(perpendicular to a line or an arm, or into a room's own hollow interior).
Searched exhaustively across five seeds with no hit.

Proven to fail two more ways: the block RNG's xorshift using a signed >>
instead of the unsigned >>> still produces *a* deterministic world, just
the wrong one — caught only by the bit-exact sweep, not by any property
test. Moving the push-out's outside/inside boundary from `bestD2 > 0` to
`>= 0` routes the exact-zero case into a `1/sqrt(0)` and fails with a
literal NaN — precisely the value the buried-body case exists to reach.

No changelog entry: the C# port is not the game.

Verified: typecheck both configs, 566 tests, dotnet test 117/117.

+++ cs/src/Scrapyard.Core/EnemyAI.cs
@@ -153,7 +153,7 @@ public static class EnemyAI
             double uy = dy * inv;
 
             double reach = rad + AvoidLookahead;
-            int ahead = scenery.Overlap(p.X[d] + ux * reach, p.Y[d] + uy * reach, rad);
+            long ahead = scenery.Overlap(p.X[d] + ux * reach, p.Y[d] + uy * reach, rad);
             bool detouring = !fixated && w.Flow.Detours(p.X[d], p.Y[d], ux, uy);
 
             if ((ahead >= 0 || detouring) && !fixated &&
@@ -172,7 +172,7 @@ public static class EnemyAI
 
                 WallTangent(scenery, ahead, p.X[d], p.Y[d], ux, uy, ccw, out double tx, out double ty);
 
-                int beside = scenery.Overlap(p.X[d] + tx * reach, p.Y[d] + ty * reach, rad);
+                long beside = scenery.Overlap(p.X[d] + tx * reach, p.Y[d] + ty * reach, rad);
                 if (beside >= 0)
                 {
                     // The slide runs into a SECOND obstacle. Try sliding along that one instead,
@@ -205,7 +205,7 @@ public static class EnemyAI
     /// would otherwise divide by zero.
     /// </summary>
     private static void WallNormal(
-        IScenery scenery, int i, double x, double y, double ux, double uy,
+        IScenery scenery, long i, double x, double y, double ux, double uy,
         out double nx, out double ny)
     {
         double dx = x - scenery.PieceX(i);
@@ -224,7 +224,7 @@ public static class EnemyAI
     }
 
     private static void WallTangent(
-        IScenery scenery, int i, double x, double y, double ux, double uy, bool ccw,
+        IScenery scenery, long i, double x, double y, double ux, double uy, bool ccw,
         out double tx, out double ty)
     {
         WallNormal(scenery, i, x, y, ux, uy, out double nx, out double ny);
